fix: [SHB-123][유저] 빌드 안되는 버그 수정

+++ src/main/java/com/smart/booking/domain/store/dto/UpsertStoreDtoTeeBoxFee.java
@@ -2,13 +2,14 @@ package com.smart.booking.domain.store.dto;
 
 import com.smart.booking.domain.common.enums.TeeBoxType;
 import com.smart.booking.domain.store.value_object.WeekdayWeekendFee;
+import java.math.BigDecimal;
 import lombok.NonNull;
 
 public record UpsertStoreDtoTeeBoxFee(
     @NonNull TeeBoxType teeBoxType,
     @NonNull WeekdayWeekendFee bookingFee,
     @NonNull WeekdayWeekendFee usageFee,
-    int onSiteFee
+    BigDecimal onSiteFee
 ) {
 
 }
